Update metadata for 2.0.5
Pyomo 5.6.9 and pyutilib 6.0 have non-backwards-compatible changes
that we didn't anticipate when we released Switch 2.0.5. We updated
later versions of Switch to avoid these versions, but now conda
installs switch 2.0.5 on platforms where the older versions of
Pyomo/pyutilib are unavailable (e.g., Apple M1/M2), falsely believing
that 2.0.5 is compatible with newer versions of Pyomo/pyutilib. This
update prevents that.